third_party/blink/web_tests/platform/generic/external/wpt/dom/nodes/adoption.window-expected.txt

This is a testharness.js-based test.
PASS appendChild() and DocumentFragment with host
FAIL adoptNode() and DocumentFragment with host assert_equals: expected Document node with 0 children but got Document node with 2 children
PASS appendChild() and DocumentFragment
PASS adoptNode() and DocumentFragment
PASS appendChild() and ShadowRoot
PASS adoptNode() and ShadowRoot
Harness: the test ran to completion.
